Refactor timing functions and clean up Application.cpp
Updated GetPhysicsFrameTime and GetRenderFrameTime to cast frame times to float. Added new timing functions (GetTicks, GetTicksMilli, GetTicksNano) and delay functions (DelayProcess, DelayProcessMilli) in Application.cpp and Application.h. Refactored Run function to use new timing and delay functions. Corrected indentation of DN_CORE_FATAL log message. Removed commented-out code related to window configuration, rendering, ImGui, and rlImGui. Reintroduced physics and rendering loops in Run function with new timing functions. Updated ECS.cpp to cast delta to float in ExecuteQuery functions.

+++ ExampleProjects/DuinFPS/src/ECS.cpp
@@ -225,7 +225,7 @@ void ExecuteQueryIdle(flecs::world& world)
             duin::Vector3 targetVel = duin::Vector3Zero();
 
             float smoothFactor = 20.75f;
-            float alpha = 1.0f - std::expf(-smoothFactor * delta);
+            float alpha = 1.0f - std::expf(-smoothFactor * (float)delta);
 
             float outputVelX = targetVel.x - velocity.value.x;
             float outputVelZ = targetVel.z - velocity.value.z;
@@ -289,7 +289,7 @@ void ExecuteQueryRun(flecs::world& world)
                duin::Vector3 targetVel = duin::Vector3Scale(iDir.value, targetSpeed);
 
                float smoothFactor = 10.75f;
-               float alpha = 1.0f - std::expf(-smoothFactor * delta);
+               float alpha = 1.0f - std::expf(-smoothFactor * (float)delta);
 
                float outputVelX = targetVel.x - velocity.value.x;
                float outputVelZ = targetVel.z - velocity.value.z;
@@ -340,7 +340,7 @@ void ExecuteQuerySprint(flecs::world& world)
                 if (duin::Vector3LengthSqr(velocity.value) < (runSpeed.speed * runSpeed.speed)) {
                     smoothFactor = 10.75f;
                 }
-                float alpha = 1.0f - std::expf(-smoothFactor * delta);
+                float alpha = 1.0f - std::expf(-smoothFactor * (float)delta);
 
                 float outputVelX = targetVel.x - velocity.value.x;
                 float outputVelZ = targetVel.z - velocity.value.z;
